CE-4860 Changed hitc to Double data type (#126)

+++ src/main/java/gov/epa/ccte/api/bioactivity/projection/assay/chemicals/CcdAssayDetails.java
@@ -84,6 +84,6 @@ public interface CcdAssayDetails {
 
 	Double getLogAc50();
 
-	Integer getHitc();
+	Double getHitc();
 
 }
